small fix on delete admin

+++ admin/delete-admin.php
@@ -1,24 +1,20 @@
-<?php 
-
+<?php
 include("../config/constants.php");
+include('partials/login-check.php');
 
-$id = $_GET['id'];
+if (isset($_SESSION['user'])) {
+    $id = $_GET['id'];
 
-$sql = "DELETE FROM tbl_admin WHERE id = $id";
+    $sql = "DELETE FROM tbl_admin WHERE id = $id";
 
-$res = mysqli_query($conn, $sql);
+    $res = mysqli_query($conn, $sql);
 
-if($res==true)
-{
-    // echo "admin deleted";
-    $_SESSION['delete'] = "<div class='success'>Admin Deleted Successfully</div>";
-    header('location:'.SITEURL.'admin/manage-admin.php');
+    if ($res == true) {
+        $_SESSION['delete'] = "<div class='success'>Admin Deleted Successfully</div>";
+        header('location:' . SITEURL . 'admin/manage-admin.php');
+    } else {
+        $_SESSION['delete'] = "<div class='error'>Failed to delete admin</div>";
+        header('location:' . SITEURL . 'admin/manage-admin.php');
+    }
 }
-else
-{
-    // echo "failed to delete admin";
-    $_SESSION['delete'] = "<div class='error'>Failed to delete admin</div>";
-    header('location:'.SITEURL.'admin/manage-admin.php');
-}
-
 ?>
